requirements.txt
streamlit
pandas
scikit-learn
langchain
langchain-core
langchain-community
langchain-openai
pinecone
python-dotenv
plotly
xgboost
seaborn
matplotlib
imbalanced-learn